+++ app/src/main/java/com/sundaydev/movieTest/data/Response.kt
@@ -1,6 +1,7 @@
 package com.sundaydev.movieTest.data
 
 import android.os.Parcelable
+import androidx.recyclerview.widget.DiffUtil
 import kotlinx.android.parcel.Parcelize
 import java.util.*
 
@@ -23,6 +24,13 @@ data class Tv(
     val vote_count: Int = 0
 ) {
     fun displayVote() = (vote_average * 10).toInt()
+
+    companion object {
+        val diffTvUtil = object : DiffUtil.ItemCallback<Tv>() {
+            override fun areItemsTheSame(oldItem: Tv, newItem: Tv): Boolean = oldItem.id == newItem.id
+            override fun areContentsTheSame(oldItem: Tv, newItem: Tv): Boolean = oldItem == newItem
+        }
+    }
 }
 
 data class Movie(
@@ -45,6 +53,13 @@ data class Movie(
         title = title, video = video, poster_path = poster_path, popularity = popularity, original_language = original_language,
         backdrop_path = backdrop_path, vote_average = vote_average, overview = overview, vote_count = vote_count
     )
+
+    companion object {
+        val diffMovieUtil = object : DiffUtil.ItemCallback<Movie>() {
+            override fun areItemsTheSame(oldItem: Movie, newItem: Movie): Boolean = oldItem.id == newItem.id
+            override fun areContentsTheSame(oldItem: Movie, newItem: Movie): Boolean = oldItem == newItem
+        }
+    }
 }
 
 data class People(
@@ -52,6 +67,13 @@ data class People(
     /*val known_for: List<String>, */val name: String, val popularity: Float
 ) {
     fun toPeopleDetail() = PeopleDetail(id = id, adult = adult, name = name, popularity = popularity, profile_path = profile_path)
+
+    companion object {
+        val diffPeopleUtil = object : DiffUtil.ItemCallback<People>() {
+            override fun areItemsTheSame(oldItem: People, newItem: People): Boolean = oldItem.id == newItem.id
+            override fun areContentsTheSame(oldItem: People, newItem: People): Boolean = oldItem == newItem
+        }
+    }
 }
 
 @Parcelize
